Split off ValidateHostHandler to a separate file
